Add SSL_VERIFY toggle, search
Toggle ssl_verify off during ncrc operation (root signed certs). Hopefully we can redact this
in the future.

Add a search function. Doesn't quite work with our setup, as the application _is_ the channel.
Meaning, you have to nail it 100% (no wild cards). In the future, we may be able to fine grain
the permissions on the server end using LDAP. That way, the channel will be universal to all
apps.